+++ cmd/forklift/plt/repositories.go
@@ -12,7 +12,7 @@ import (
 
 func cacheRepoAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, false)
+		pallet, cache, err := processFullBaseArgs(c.String("workspace"), false)
 		if err != nil {
 			return err
 		}
@@ -35,7 +35,7 @@ func cacheRepoAction(versions Versions) cli.ActionFunc {
 
 		// TODO: warn if any downloaded repo doesn't appear to be an actual repo, or if any repo's
 		// forklift version is incompatible or ahead of the pallet version
-		fmt.Println("Done! Next, you'll probably want to run `sudo -E forklift plt apply`.")
+		fmt.Println("Done!")
 		return nil
 	}
 }
@@ -51,14 +51,42 @@ func lsRepoAction(c *cli.Context) error {
 	return fcli.PrintPalletRepos(0, pallet)
 }
 
-// show-plt
+// show-repo
 
 func showRepoAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
 	if err != nil {
 		return err
 	}
 
-	repoPath := c.Args().First()
-	return fcli.PrintRepoInfo(0, pallet, cache, repoPath)
+	return fcli.PrintRepoInfo(0, pallet, cache, c.Args().First())
+}
+
+// add-repo
+
+func addRepoAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		pallet, repoCache, err := processFullBaseArgs(c.String("workspace"), false)
+		if err != nil {
+			return err
+		}
+		if err = fcli.CheckShallowCompatibility(
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+
+		if err = fcli.AddRepoRequirements(0, pallet, repoCache.Path(), c.Args().Slice()); err != nil {
+			return err
+		}
+
+		if !c.Bool("no-cache-req") {
+			if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
+				return err
+			}
+		}
+		fmt.Println("Done!")
+		return nil
+	}
 }